\name{pnnclass-package}
\alias{pnnclass-package}
\alias{pnnclass}
\docType{package}
\title{
\packageTitle{pnnclass}
}
\description{
\packageDescription{pnnclass}
}
% \details{
%   This section should provide a more detailed overview of how to use the
%   package, including the most important functions.
% }
\author{
  \packageAuthor{pnnclass}
  
  Maintainer: \packageMaintainer{pnnclass}
}
% \references{
% }
% \keyword{package}
% \seealso{
%   Optional links to other man pages
% }
% \examples{
%   \dontrun{
%      ## Optional simple examples of the most important functions
%      ## These can be in \dontrun{} and \donttest{} blocks.   
%   }
% }
